+++ rl_blox/blox/function_approximator/mlp.py
@@ -1,0 +1,74 @@
+from typing import Callable
+
+import chex
+import jax.numpy as jnp
+from flax import nnx
+
+
+class MLP(nnx.Module):
+    """Multilayer Perceptron.
+
+    Parameters
+    ----------
+    n_features : int
+        Number of features.
+
+    n_outputs : int
+        Number of output components.
+
+    hidden_nodes : list
+        Numbers of hidden nodes of the MLP.
+
+    activation : str
+        Activation function. Has to be the name of a function defined in the
+        flax.nnx module.
+
+    rngs nnx.Rngs
+        Random number generator.
+
+    Attributes
+    ----------
+    n_outputs : int
+        Number of output components.
+
+    activation : Callable
+        Activation function.
+
+    hidden_layers : list[nnx.Linear]
+        Hidden layers.
+
+    output_layer : nnx.Linear
+        Output layer.
+    """
+
+    n_outputs: int
+    activation: Callable[[jnp.ndarray], jnp.ndarray]
+    hidden_layers: list[nnx.Linear]
+    output_layer: nnx.Linear
+
+    def __init__(
+        self,
+        n_features: int,
+        n_outputs: int,
+        hidden_nodes: list[int],
+        activation: str,
+        rngs: nnx.Rngs,
+    ):
+        chex.assert_scalar_positive(n_features)
+        chex.assert_scalar_positive(n_outputs)
+
+        self.n_outputs = n_outputs
+        self.activation = getattr(nnx, activation)
+
+        self.hidden_layers = []
+        n_in = n_features
+        for n_out in hidden_nodes:
+            self.hidden_layers.append(nnx.Linear(n_in, n_out, rngs=rngs))
+            n_in = n_out
+
+        self.output_layer = nnx.Linear(n_in, n_outputs, rngs=rngs)
+
+    def __call__(self, x: jnp.ndarray) -> jnp.ndarray:
+        for layer in self.hidden_layers:
+            x = self.activation(layer(x))
+        return self.output_layer(x)
